+++ src/location/track/mod.rs
@@ -9,7 +9,7 @@ pub struct BalloonTrack {
 }
 
 impl BalloonTrack {
-    pub fn new(name: String, callsign: Option<String>) -> Self {
+    pub fn new(name: String) -> Self {
         Self {
             locations: vec![],
             prediction: None,
@@ -37,67 +37,6 @@ impl BalloonTrack {
         false
     }
 
-    pub fn intervals(&self) -> Vec<chrono::Duration> {
-        let mut values = vec![];
-
-        for index in 0..(self.locations.len() - 1) {
-            let current = self.locations.get(index).unwrap();
-            let next = self.locations.get(index + 1).unwrap();
-            values.push(next.time - current.time);
-        }
-
-        values
-    }
-
-    pub fn ascents(&self) -> Vec<f64> {
-        let mut values = vec![];
-
-        for index in 0..(self.locations.len() - 1) {
-            let current = self.locations.get(index).unwrap();
-            let next = self.locations.get(index + 1).unwrap();
-            values.push(
-                next.altitude.expect("location has no altitude")
-                    - current.altitude.expect("location has no altitude"),
-            );
-        }
-
-        values
-    }
-
-    pub fn ascent_rates(&self) -> Vec<f64> {
-        let mut values = vec![];
-        let interval = self.intervals();
-
-        for (index, ascent) in self.ascents().iter().enumerate() {
-            values.push(ascent / interval.get(index).unwrap().num_seconds() as f64);
-        }
-
-        values
-    }
-
-    pub fn overground_distances(&self) -> Vec<f64> {
-        let mut values = vec![];
-
-        for index in 0..(self.locations.len() - 1) {
-            let current = self.locations.get(index).unwrap();
-            let next = self.locations.get(index + 1).unwrap();
-            values.push(current.location.geodesic_distance(&next.location));
-        }
-
-        values
-    }
-
-    pub fn ground_speeds(&self) -> Vec<f64> {
-        let mut values = vec![];
-        let interval = self.intervals();
-
-        for (index, distance) in self.overground_distances().iter().enumerate() {
-            values.push(distance / interval.get(index).unwrap().num_seconds() as f64);
-        }
-
-        values
-    }
-
     pub fn estimated_time_to_ground(&self) -> Option<chrono::Duration> {
         if !self.locations.is_empty() && self.descending() {
             if let Some(freefall_estimate) = self.falling() {
@@ -111,7 +50,8 @@ impl BalloonTrack {
                 }
                 if altitudes.len() > 1 {
                     Some(chrono::Duration::milliseconds(
-                        ((-self.ascent_rates().last().unwrap() / altitudes.last().unwrap())
+                        ((-ascent_rates(&self.locations).last().unwrap()
+                            / altitudes.last().unwrap())
                             * 1000.0) as i64,
                     ))
                 } else {
@@ -132,12 +72,12 @@ impl BalloonTrack {
     }
 
     pub fn ascending(&self) -> bool {
-        let ascent_rates = self.ascent_rates();
+        let ascent_rates = ascent_rates(&self.locations);
         ascent_rates.iter().rev().take(2).all(|a| a > &0.2)
     }
 
     pub fn descending(&self) -> bool {
-        let ascent_rates = self.ascent_rates();
+        let ascent_rates = ascent_rates(&self.locations);
         ascent_rates.iter().rev().take(2).all(|a| a < &0.2)
     }
 
@@ -147,7 +87,7 @@ impl BalloonTrack {
         if last_location.altitude.is_some() && self.descending() {
             let freefall_estimate = last_location.estimate_freefall();
 
-            if let Some(last_ascent_rate) = self.ascent_rates().last() {
+            if let Some(last_ascent_rate) = ascent_rates(&self.locations).last() {
                 if (last_ascent_rate - freefall_estimate.ascent_rate)
                     < freefall_estimate.ascent_rate_uncertainty
                 {
@@ -177,4 +117,100 @@ impl BalloonTrackAttributes {
     pub fn new() -> Self {
         Self { callsign: None }
     }
+}
+
+pub fn with_altitude(locations: &Vec<super::BalloonLocation>) -> Vec<super::BalloonLocation> {
+    locations
+        .into_iter()
+        .filter(|location| location.altitude.is_some())
+        .map(|location| location.to_owned())
+        .collect()
+}
+
+pub fn intervals(locations: &Vec<super::BalloonLocation>) -> Vec<chrono::Duration> {
+    let mut values = vec![];
+
+    for index in 0..(locations.len() - 1) {
+        let current = locations.get(index).unwrap();
+        let next = locations.get(index + 1).unwrap();
+        values.push(next.time - current.time);
+    }
+
+    values
+}
+
+pub fn altitudes(locations: &Vec<super::BalloonLocation>) -> Vec<f64> {
+    locations
+        .iter()
+        .filter_map(|locations| locations.altitude)
+        .collect()
+}
+
+pub fn ascents(locations: &Vec<super::BalloonLocation>) -> Vec<f64> {
+    let mut values = vec![];
+
+    let mut index = 0;
+    let mut current = locations.first().unwrap();
+    let mut next;
+    loop {
+        next = match locations.get(index + 1) {
+            Some(next) => next,
+            None => {
+                break;
+            }
+        };
+
+        values.push(next.altitude.unwrap() - current.altitude.unwrap());
+
+        current = next;
+        index += 1;
+    }
+
+    values
+}
+
+pub fn ascent_rates(locations: &Vec<super::BalloonLocation>) -> Vec<f64> {
+    let mut values = vec![];
+
+    let locations_with_altitude = with_altitude(locations);
+    let intervals = intervals(&locations_with_altitude);
+    for (index, ascent) in ascents(&locations_with_altitude).iter().enumerate() {
+        values.push(ascent / intervals.get(index).unwrap().num_seconds() as f64);
+    }
+
+    values
+}
+
+pub fn overground_distances(locations: &Vec<super::BalloonLocation>) -> Vec<f64> {
+    let mut values = vec![];
+
+    let mut index = 0;
+    let mut current = locations.first().unwrap();
+    let mut next;
+    loop {
+        next = match locations.get(index + 1) {
+            Some(next) => next,
+            None => {
+                break;
+            }
+        };
+
+        values.push(current.location.geodesic_distance(&next.location));
+
+        current = next;
+        index += 1;
+    }
+
+    values
+}
+
+pub fn ground_speeds(locations: &Vec<super::BalloonLocation>) -> Vec<f64> {
+    let mut values = vec![];
+
+    let intervals = intervals(locations);
+    for (index, distance) in overground_distances(locations).iter().enumerate() {
+        values.push(distance / intervals.get(index).unwrap().num_seconds() as f64);
+    }
+
+    values
 }
